moved debug tools
