+++ stwo_cairo_prover/crates/prover/src/cairo_air/mod.rs
@@ -6,7 +6,6 @@ pub(crate) mod air;
 pub(crate) mod blake;
 pub(crate) mod builtins_air;
 pub mod components;
-pub(crate) mod debug_tools;
 pub(crate) mod opcodes_air;
 pub(crate) mod pedersen;
 pub(crate) mod poseidon;
