+++ stwo_cairo_prover/crates/prover/src/cairo_air/range_checks_air.rs
@@ -8,11 +8,11 @@ use stwo_prover::core::channel::Channel;
 use stwo_prover::core::fields::qm31::{SecureField, QM31};
 use stwo_prover::core::pcs::TreeVec;
 
-use super::debug_tools::indented_component_display;
 use crate::cairo_air::components::{
-    range_check_11, range_check_12, range_check_18, range_check_19, range_check_3_3_3_3_3,
-    range_check_3_6, range_check_3_6_6_3, range_check_4_3, range_check_4_4, range_check_4_4_4_4,
-    range_check_5_4, range_check_6, range_check_7_2_5, range_check_8, range_check_9_9,
+    indented_component_display, range_check_11, range_check_12, range_check_18, range_check_19,
+    range_check_3_3_3_3_3, range_check_3_6, range_check_3_6_6_3, range_check_4_3, range_check_4_4,
+    range_check_4_4_4_4, range_check_5_4, range_check_6, range_check_7_2_5, range_check_8,
+    range_check_9_9,
 };
 use crate::cairo_air::relations;
