+++ stwo_cairo_prover/crates/prover/src/debug_tools/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(test)]
+pub mod assert_constraints;
+
+#[cfg(test)]
+pub mod mock_tree_builder;
+
+#[cfg(feature = "relation-tracker")]
+pub mod relation_tracker;
